+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/queue/ProcessCompletionCondition.java
@@ -24,6 +24,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import org.immutables.value.Value;
 
+import javax.annotation.Nullable;
 import java.util.List;
 import java.util.UUID;
 
@@ -31,6 +32,9 @@ import java.util.UUID;
 @JsonSerialize(as = ImmutableProcessCompletionCondition.class)
 @JsonDeserialize(as = ImmutableProcessCompletionCondition.class)
 public abstract class ProcessCompletionCondition extends AbstractWaitCondition {
+
+    @Nullable
+    public abstract String resumeEvent();
 
     public abstract List<UUID> processes();