+++ server/impl/src/test/java/com/walmartlabs/concord/server/process/queue/WaitConditionSerializeTest.java
@@ -65,7 +65,9 @@ public class WaitConditionSerializeTest {
 
     @Test
     public void testProcessCompletionCondition() throws Exception {
-        ProcessCompletionCondition c = ProcessCompletionCondition.of(Collections.singletonList(UUID.fromString("0c443946-0f2c-4685-a9f0-2bc0b735b7ae")),"test-reason");
+        ProcessCompletionCondition c = ProcessCompletionCondition.of(
+                Collections.singletonList(UUID.fromString("0c443946-0f2c-4685-a9f0-2bc0b735b7ae")),
+                "test-reason");
 
         String json = objectMapper.writeValueAsString(c);
         System.out.println(json);